backend/requirements.txt
fastapi
uvicorn
google-generativeai
langchain
langchain-google-genai
chromadb
python-multipart
pymupdf
python-pptx
pydantic
python-dotenv
requests
